[MK] Changed variables from float to double

+++ app/src/main/java/com/example/c0777180_w2020_mad3125_midterm/DataDisplayActivity.java
@@ -52,10 +52,11 @@ public class DataDisplayActivity extends AppCompatActivity {
         String c = personInformation.getGender();
         int d = personInformation.getAge();
         String e = personInformation.getDateOfBirth();
-        float f = personInformation.getGrossIncome();
-        float g = personInformation.getRRSPContributed();
+        double f = personInformation.getGrossIncome();
+        double g = personInformation.getRRSPContributed();
         String h = personInformation.getTaxFilingDate();
-        float j = personInformation.getCPP();
+        double j = personInformation.getCPP();
+
 
         SIN.setText(a);
         fullname.setText(b);
